normalize gas price

+++ nextjs-app/app/api/chat/route.ts
@@ -36,7 +36,7 @@ interface AIResponse {
 const schema = `
 CREATE TABLE ethereum_blocks (timestamp bigint, number bigint, hash character(66), extradata text, size bigint, gasused numeric);
 
-CREATE TABLE ethereum_transactions (timestamp bigint, hash character(66), fromaddr character(42), toaddr character(42), value numeric, gas bigint, gasprice bigint, input text, blocknumber bigint);
+CREATE TABLE ethereum_transactions (timestamp bigint, hash character(66), fromaddr character(42), toaddr character(42), value numeric, gas bigint, gasprice numeric, input text, blocknumber bigint);
 
 CREATE TABLE fetch_erc20_transfers (timestamp bigint, fromaddr character(42), toaddr character(42), value numeric, transactionhash character(66), logindex integer, blocknumber bigint);
